Moved register-hours modal and date bugfix 🐛
