+++ ClientApp/src/app/components/modals/register-hours/register-hours-modal.component.html
@@ -17,8 +17,7 @@
     ></dp-input>
 
     <dp-autocomplete
-      [options]="options"
-      [value]="{ name: 'asdf' }"
+      [options]="projectOptions"
       optionField="name"
       label="Select Project"
       formControlName="project"
