+++ ClientApp/src/app/components/modals/register-hours/register-hours-modal.component.scss
