+++ ClientApp/src/app/components/modals/register-hours/register-hours-modal.component.ts
@@ -7,16 +7,16 @@ import {
   AbstractControl,
 } from '@angular/forms';
 
-import { UserService } from '../../services/user.service';
-import { HoursService } from '../../services/hours.service';
+import { UserService } from '../../../services/user.service';
+import { HoursService } from '../../../services/hours.service';
 
 import { User } from 'src/app/models/user.model';
 import { Hours } from 'src/app/models/hours.model';
 
 @Component({
-  selector: 'app-register-hours',
-  templateUrl: './register-hours.component.html',
-  styleUrls: ['./register-hours.component.scss'],
+  selector: 'app-register-hours-modal',
+  templateUrl: './register-hours-modal.component.html',
+  styleUrls: ['./register-hours-modal.component.scss'],
 })
 export class RegisterHoursComponent implements OnInit {
   constructor(
@@ -41,13 +41,9 @@ export class RegisterHoursComponent implements OnInit {
   });
 
   // TODO: Replace with data from DB
-  options = ['Tempus', 'Summer Internship', 'Ship GUI'].map((x) => {
+  projectOptions = ['Tempus', 'Summer Internship', 'Ship GUI'].map((x) => {
     return { name: x };
   });
-
-  selectedProject = this.options[0];
-
-  jsonString: string;
 
   ngOnInit(): void {
     this.updateHoursKey();
@@ -58,8 +54,8 @@ export class RegisterHoursComponent implements OnInit {
     return (n < 10 ? '0' + n : n).toString();
   }
 
+  // Updates the hourkey to the selected date zero-padded
   updateHoursKey(): void {
-    // Updates the hourkey to the selected date zero-padded
     const day = this.zeroPad(this.selectedDate.getDate());
     const month = this.zeroPad(this.selectedDate.getMonth() + 1);
     const year = this.selectedDate.getFullYear();
@@ -69,14 +65,20 @@ export class RegisterHoursComponent implements OnInit {
   updateDate(date: Date): void {
     this.selectedDate = date;
     this.updateHoursKey();
-    this.getUserData();
+    this.updateForm();
+
+    // Update the date in the form to UTC of date
+    this.hoursRegisterForm.patchValue({
+      date: date.getTime(),
+    });
   }
 
+  // Get the user data and update the form
   getUserData() {
-    // Get the user data and update the form
     this.userService.getCurrentUser().then((resp) => {
       this.user = resp;
       this.updateForm();
+      console.log(this.user.Hours[this.hoursKey]);
     });
   }
 
@@ -85,33 +87,36 @@ export class RegisterHoursComponent implements OnInit {
     this.hoursRegisterForm.patchValue({
       project: this.hoursRegisterForm.get('project').value.name,
     });
-    this.jsonString = JSON.stringify(this.hoursRegisterForm.value);
     const hours: Hours = this.hoursRegisterForm.value;
     const id = this.user._id;
     this.hoursService.registerHours(id, hours).then((resp) => {
       if (resp) {
-        location.reload();
-        this.router.navigate(['/']);
+        this.closeModal();
+        /* console.log(this.hoursRegisterForm.value); */
+        /* location.reload(); */
+        /* this.router.navigate(['/']); */
+      } else {
       }
     });
   }
 
   updateForm() {
+    // Find the project in the hours object
     let project: {} = this.getHoursObject().Project;
     if (project === null || project === undefined) {
-      // No project found, use first project in dropdown
-      project = this.options[0].name;
+      // The hour object didn't have a project field, so use first value in project dropdown
+      project = this.projectOptions[0].name;
     }
+
     // Patch the form hours and project values to the values from the database
     this.hoursRegisterForm.patchValue({
       hours: this.getHoursObject().Hours,
       project: { name: project },
     });
-    /* this.hoursDayForm.controls.project.setValidators(this.validHourValidator()); */
   }
 
+  // Return the wanted hours object from the database or an empty object
   getHoursObject(): Hours {
-    // Return the wanted hours object from the database or an empty object
     return this.user.Hours[this.hoursKey] || {};
   }
